feat: bootloader now runs and loads kernel

+++ src/kernel/paging.cxx
@@ -2,8 +2,8 @@
 #include "kernel/asmfuncs.hxx"
 #include "kernel/mem.hxx"
 
-using PageDirectoryEntry = uint32_t;
-using PageTableEntry = uint32_t;
+using PageDirectoryEntry = uint64_t;
+using PageTableEntry = uint64_t;
 
 volatile PageDirectoryEntry pageDirectory[1024] __attribute__ ((aligned (4096)));
 volatile PageTableEntry firstPageTable[1024] __attribute__ ((aligned (4096)));
@@ -12,6 +12,6 @@ void beginPaging() {
 	for (auto i = 0; i < 1024; i++) {
 		firstPageTable[i] = (i * 0x1000) | 3;
 	}
-	pageDirectory[0] = ((unsigned int) firstPageTable) | 3;
+	pageDirectory[0] = ((uint64_t)firstPageTable) | 3;
 	setupPaging((void*)&pageDirectory);
 }
